Add files via upload
The task is to cut the whole board(matrix) resulting in splitting of every block. Given the cost of cutting from row side(horizontal) and column side(vertical). Find the minimal cost of cutting the whole board into pieces. 
NOTE: If a row or column cut goes through a block that is already been detached, the cost is multiplied by number of separate blocks.
GREEDY APPROACH.